PrintList

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,7 +8,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	size_t number;
+	int number;
 
 	number = 0;
 	while (h != NULL)
